New Docker setup based on docker-sync
This is a large update which includes many (breaking) changes. The new setup involves docker-sync for near-native performance and for bi-directional file sync capability. This setup also includes xdebug for debugging.

The images are generally multi-use, which means you can use them on either Magento 1 or Magento 2, perhaps with just minor changes.Updated README to follow.